+++ python_modules/dagster/dagster/_utils/caching_instance_queryer.py
@@ -43,12 +43,12 @@ from dagster._core.errors import (
 from dagster._core.event_api import AssetRecordsFilter, EventRecordsFilter
 from dagster._core.events import DagsterEventType
 from dagster._core.instance import DagsterInstance, DynamicPartitionsStore
+from dagster._core.storage.batch_asset_record_loader import BatchAssetRecordLoader
 from dagster._core.storage.dagster_run import (
     DagsterRun,
     RunRecord,
 )
 from dagster._core.storage.tags import PARTITION_NAME_TAG
-from dagster._core.workspace.batch_asset_record_loader import BatchAssetRecordLoader
 from dagster._utils.cached_method import cached_method
 
 if TYPE_CHECKING:
