+++ python_modules/dagster/dagster_tests/storage_tests/utils/event_log_storage.py
@@ -3690,7 +3690,7 @@ class TestEventLogStorage:
                     )
                 )[0]
 
-                if storage.asset_records_have_last_planned_materialization_storage_id():
+                if storage.asset_records_have_last_planned_materialization_storage_id:
                     assert (
                         asset_entry.last_planned_materialization_storage_id
                         == materialization_planned_record.storage_id
